requirements.txt
pyTelegramBotAPI
python-dotenv
openai
requests
